pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "dumpsleuth"
version = "2.0.0"
description = "Modern Memory Dump Analysis Toolkit"
readme = "README.md"
license = {file = "LICENSE"}
authors = [{name = "James Nelson", email = "james.nelson@jamesnelsonsec.xyz"}]
maintainers = [{name = "James Nelson", email = "james.nelson@jamesnelsonsec.xyz"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
requires-python = ">=3.7"
dependencies = [
    "click>=8.0.0",
    "rich>=10.0.0",
    "pyyaml>=5.4",
    "python-magic-bin>=0.4.14;platform_system=='Windows'",
    "python-magic>=0.4.24;platform_system!='Windows'",
    "hexdump>=3.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.9",
    "mypy>=0.910",
    "sphinx>=4.0",
]
web = [
    "fastapi>=0.68",
    "uvicorn>=0.15",
    "jinja2>=3.0",
]

[project.urls]
Homepage = "https://github.com/OVHGERMANY/DumpSleuth"
"Bug Tracker" = "https://github.com/OVHGERMANY/DumpSleuth/issues"
Documentation = "https://github.com/OVHGERMANY/DumpSleuth/wiki"
Repository = "https://github.com/OVHGERMANY/DumpSleuth"

[project.scripts]
dumpsleuth = "dumpsleuth.ui.cli:cli"
ds = "dumpsleuth.ui.cli:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
dumpsleuth = [
    "config/*.yaml",
    "templates/*.html",
    "templates/*.css", 
    "templates/*.js",
]

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=dumpsleuth",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-branch",
]

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["magic.*", "hexdump.*"]
ignore_missing_imports = true 